refac: OrderController
