+++ src/main/java/HighThroughPutExchange/API/ServerApplication.java
@@ -13,7 +13,6 @@ import HighThroughPutExchange.Auction.Auction;
 import HighThroughPutExchange.Common.Message;
 import HighThroughPutExchange.Common.TaskFuture;
 import HighThroughPutExchange.Common.TaskQueue;
-import HighThroughPutExchange.API.service.OrderService;
 import HighThroughPutExchange.Database.entry.DBEntry;
 import HighThroughPutExchange.Database.exceptions.AlreadyExistsException;
 import HighThroughPutExchange.Database.exceptions.NotFoundException;
@@ -65,7 +64,6 @@ public class ServerApplication {
     private static final int KEY_LENGTH = 16;
     private static final int MAX_OPERATIONS = 20;
     private MatchingEngine matchingEngine;
-    private OrderService orderService;
 
     private static char randomChar() {
         return (char) ((int) (Math.random() * 26 + 65));
@@ -79,9 +77,9 @@ public class ServerApplication {
         return output.toString();
     }
 
-    public ServerApplication(MatchingEngine matchingEngine, OrderService orderService) {
+    public ServerApplication(MatchingEngine matchingEngine, RateLimiter rateLimiter) {
         this.matchingEngine = matchingEngine;
-        this.orderService = orderService;
+        this.rateLimiter = rateLimiter;
         state = State.STOP;
         HashMap<String, Class<? extends DBEntry>> mapping = new HashMap<>();
         mapping.put("users", User.class);
@@ -147,12 +145,15 @@ public class ServerApplication {
         privatePageAuthenticator = PrivatePageAuthenticator.getInstance();
         BotAuthenticator.buildInstance(botSessions);
         botAuthenticator = BotAuthenticator.getInstance();
-        rateLimiter = new RateLimiter();
         auction = new Auction(matchingEngine);
     }
 
     public static void main(String[] args) {
         SpringApplication.run(ServerApplication.class, args);
+    }
+
+    public State getState() {
+        return state;
     }
 
     /*
@@ -464,132 +465,6 @@ public class ServerApplication {
     }
 
     @CrossOrigin(origins = "*")
-    @PostMapping("/limit_order")
-    public ResponseEntity<LimitOrderResponse> limitOrder(@Valid @RequestBody LimitOrderRequest form) {
-        if (!privatePageAuthenticator.authenticate(form)) {
-            return new ResponseEntity<>(new LimitOrderResponse(Message.AUTHENTICATION_FAILED.toString()), HttpStatus.UNAUTHORIZED);
-        }
-        if (!rateLimiter.processRequest(form)) {
-            return new ResponseEntity<>(new LimitOrderResponse(Message.RATE_LIMITED.toString()), HttpStatus.TOO_MANY_REQUESTS);
-        }
-        if (state == State.STOP) {
-            return new ResponseEntity<>(new LimitOrderResponse(Message.TRADE_LOCKED.toString()), HttpStatus.LOCKED);
-        }
-        String message = orderService.placeLimitOrder(form.getUsername(), form.getTicker(), form.getPrice(), form.getVolume(), form.getBid());
-        return new ResponseEntity<>(new LimitOrderResponse(message), HttpStatus.OK);
-    }
-
-    @CrossOrigin(origins = "*")
-    @PostMapping("/bot_limit_order")
-    public ResponseEntity<LimitOrderResponse> botLimitOrder(@Valid @RequestBody BotLimitOrderRequest form) {
-        if (!botAuthenticator.authenticate(form)) {
-            return new ResponseEntity<>(new LimitOrderResponse(Message.AUTHENTICATION_FAILED.toString()), HttpStatus.UNAUTHORIZED);
-        }
-        if (state == State.STOP) {
-            return new ResponseEntity<>(new LimitOrderResponse(Message.TRADE_LOCKED.toString()), HttpStatus.LOCKED);
-        }
-        String message = orderService.placeLimitOrder(form.getUsername(), form.getTicker(), form.getPrice(), form.getVolume(), form.getBid());
-        return new ResponseEntity<>(new LimitOrderResponse(message), HttpStatus.OK);
-    }
-
-    @CrossOrigin(origins = "*")
-    @PostMapping("/remove_all")
-    public ResponseEntity<RemoveAllResponse> removeAll(@Valid @RequestBody RemoveAllRequest form) {
-        if (!privatePageAuthenticator.authenticate(form)) {
-            return new ResponseEntity<>(new RemoveAllResponse(Message.AUTHENTICATION_FAILED.toString()), HttpStatus.UNAUTHORIZED);
-        }
-        if (!rateLimiter.processRequest(form)) {
-            return new ResponseEntity<>(new RemoveAllResponse(Message.RATE_LIMITED.toString()), HttpStatus.TOO_MANY_REQUESTS);
-        }
-        if (state == State.STOP) {
-            return new ResponseEntity<>(new RemoveAllResponse(Message.TRADE_LOCKED.toString()), HttpStatus.LOCKED);
-        }
-        if (form.getUsername() == null)
-            return new ResponseEntity<>(new RemoveAllResponse(Message.AUTHENTICATION_FAILED.toString()), HttpStatus.UNAUTHORIZED);
-        String message = orderService.removeAll(form.getUsername());
-        return new ResponseEntity<>(new RemoveAllResponse(message), HttpStatus.OK);
-    }
-
-    @PostMapping("/bot_remove_all")
-    public ResponseEntity<RemoveAllResponse> botRemoveAll(@Valid @RequestBody RemoveAllRequest form) {
-        if (!botAuthenticator.authenticate(form)) {
-            return new ResponseEntity<>(new RemoveAllResponse(Message.AUTHENTICATION_FAILED.toString()), HttpStatus.UNAUTHORIZED);
-        }
-        if (!rateLimiter.processRequest(form)) {
-            return new ResponseEntity<>(new RemoveAllResponse(Message.RATE_LIMITED.toString()), HttpStatus.TOO_MANY_REQUESTS);
-        }
-        if (state == State.STOP) {
-            return new ResponseEntity<>(new RemoveAllResponse(Message.TRADE_LOCKED.toString()), HttpStatus.LOCKED);
-        }
-        if (form.getUsername() == null)
-            return new ResponseEntity<>(new RemoveAllResponse(Message.AUTHENTICATION_FAILED.toString()), HttpStatus.UNAUTHORIZED);
-        String message = orderService.removeAll(form.getUsername());
-        return new ResponseEntity<>(new RemoveAllResponse(message), HttpStatus.OK);
-    }
-
-    @CrossOrigin(origins = "*")
-    @PostMapping("/remove")
-    public ResponseEntity<RemoveAllResponse> remove(@Valid @RequestBody RemoveRequest form) {
-        if (!privatePageAuthenticator.authenticate(form)) {
-            return new ResponseEntity<>(new RemoveAllResponse(Message.AUTHENTICATION_FAILED.toString()), HttpStatus.UNAUTHORIZED);
-        }
-        if (!rateLimiter.processRequest(form)) {
-            return new ResponseEntity<>(new RemoveAllResponse(Message.RATE_LIMITED.toString()), HttpStatus.TOO_MANY_REQUESTS);
-        }
-        if (state == State.STOP) {
-            return new ResponseEntity<>(new RemoveAllResponse(Message.TRADE_LOCKED.toString()), HttpStatus.LOCKED);
-        }
-        if (form.getUsername() == null)
-            return new ResponseEntity<>(new RemoveAllResponse(Message.AUTHENTICATION_FAILED.toString()), HttpStatus.UNAUTHORIZED);
-        String message = orderService.removeOrder(form.getUsername(), form.getOrderID());
-        return new ResponseEntity<>(new RemoveAllResponse(message), HttpStatus.OK);
-    }
-
-    @CrossOrigin(origins = "*")
-    @PostMapping("/bot_remove")
-    public ResponseEntity<RemoveAllResponse> botRemove(@Valid @RequestBody RemoveRequest form) {
-        if (!botAuthenticator.authenticate(form)) {
-            return new ResponseEntity<>(new RemoveAllResponse(Message.AUTHENTICATION_FAILED.toString()), HttpStatus.UNAUTHORIZED);
-        }
-        if (state == State.STOP) {
-            return new ResponseEntity<>(new RemoveAllResponse(Message.TRADE_LOCKED.toString()), HttpStatus.LOCKED);
-        }
-        if (form.getUsername() == null)
-            return new ResponseEntity<>(new RemoveAllResponse(Message.AUTHENTICATION_FAILED.toString()), HttpStatus.UNAUTHORIZED);
-        String message = orderService.removeOrder(form.getUsername(), form.getOrderID());
-        return new ResponseEntity<>(new RemoveAllResponse(message), HttpStatus.OK);
-    }
-
-    @CrossOrigin(origins = "*")
-    @PostMapping("/market_order")
-    public ResponseEntity<MarketOrderResponse> marketOrderResponse(@Valid @RequestBody MarketOrderRequest form) {
-        if (!privatePageAuthenticator.authenticate(form)) {
-            return new ResponseEntity<>(new MarketOrderResponse(Message.AUTHENTICATION_FAILED.toString()), HttpStatus.UNAUTHORIZED);
-        }
-        if (!rateLimiter.processRequest(form)) {
-            return new ResponseEntity<>(new MarketOrderResponse(Message.RATE_LIMITED.toString()), HttpStatus.TOO_MANY_REQUESTS);
-        }
-        if (state == State.STOP) {
-            return new ResponseEntity<>(new MarketOrderResponse(Message.TRADE_LOCKED.toString()), HttpStatus.LOCKED);
-        }
-        String message = orderService.placeMarketOrder(form.getUsername(), form.getTicker(), form.getVolume(), form.getBid());
-        return new ResponseEntity<>(new MarketOrderResponse(message), HttpStatus.OK);
-    }
-
-    @CrossOrigin(origins = "*")
-    @PostMapping("/bot_market_order")
-    public ResponseEntity<MarketOrderResponse> botMarketOrderResponse(@Valid @RequestBody BotMarketOrderRequest form) {
-        if (!botAuthenticator.authenticate(form)) {
-            return new ResponseEntity<>(new MarketOrderResponse(Message.AUTHENTICATION_FAILED.toString()), HttpStatus.UNAUTHORIZED);
-        }
-        if (state == State.STOP) {
-            return new ResponseEntity<>(new MarketOrderResponse(Message.TRADE_LOCKED.toString()), HttpStatus.LOCKED);
-        }
-        String message = orderService.placeMarketOrder(form.getUsername(), form.getTicker(), form.getVolume(), form.getBid());
-        return new ResponseEntity<>(new MarketOrderResponse(message), HttpStatus.OK);
-    }
-
-    @CrossOrigin(origins = "*")
     @PostMapping("/batch")
     public ResponseEntity<BatchResponse> processBatch(@Valid @RequestBody BatchRequest form) {
         if (!botAuthenticator.authenticate(form)) {
@@ -647,7 +522,6 @@ public class ServerApplication {
                     });
                     break;
                 case "remove_all" :
-                    RemoveAllOperation removeAllOperation = (RemoveAllOperation) operation;
                     temporaryTaskQueue.add(() -> {
                         // System.out.println("Batch - Remove All");
                         matchingEngine.removeAll(form.getUsername(), future);
@@ -670,7 +544,6 @@ public class ServerApplication {
         for (int i = 0; i < futures.size(); i++) {
             TaskFuture<String> future = futures.get(i);
             future.waitForCompletion();
-            String message = future.getData();
             // System.out.printf("Batch %d Message: %s\n", i, message);
             OperationResponse response = responses.get(i);
             response.setMessage(future.getData());
